Pass POJO to component; get maps
This commit introduces new components:
interactive-layers
layer-groups

This is the first draft of a structure for abstracting away some of the mapbox-gl-source + mapbox-gl-layer combinations that were required to get a layer on the map. It introduces a new object structure, an array of objects (source) each containing layers property, an array of object (layer).